+++ tests/lzma2_mt.rs
@@ -1,18 +1,18 @@
 use std::io::{Cursor, Read, Write};
 
-use lzma_rust2::{LZMA2Options, LZMA2ReaderMT, LZMA2WriterMT, MIN_STREAM_SIZE};
+use lzma_rust2::{LZMA2ReaderMT, LZMA2WriterMT, LZMAOptions, MIN_STREAM_SIZE};
 
 static EXECUTABLE: &[u8] = include_bytes!("data/executable.exe");
 static PG100: &[u8] = include_bytes!("data/pg100.txt");
 static PG6800: &[u8] = include_bytes!("data/pg6800.txt");
 
 fn test_round_trip(data: &[u8], level: u32) {
-    let option = LZMA2Options::with_preset(level);
+    let option = LZMAOptions::with_preset(level);
 
     let mut compressed = Vec::new();
 
     {
-        let mut writer = LZMA2WriterMT::new(&mut compressed, &option, MIN_STREAM_SIZE, 4);
+        let mut writer = LZMA2WriterMT::new(&mut compressed, &option, MIN_STREAM_SIZE, 32);
         writer.write_all(data).unwrap();
         writer.finish().unwrap();
     }
@@ -20,7 +20,7 @@ fn test_round_trip(data: &[u8], level: u32) {
     let mut uncompressed = Vec::new();
 
     {
-        let mut reader = LZMA2ReaderMT::new(Cursor::new(compressed), option.dict_size, None, 2);
+        let mut reader = LZMA2ReaderMT::new(Cursor::new(compressed), option.dict_size, None, 32);
         reader.read_to_end(&mut uncompressed).unwrap();
         assert!(reader.stream_count() > 1);
     }